Implements Arvind's feedback.

+++ cleanup.sh
@@ -16,6 +16,7 @@ export BIN_DIR="${BIN_DIR:-/usr/local/bin}"
 export ISTIO_INJECT_NS="${ISTIO_INJECT_NS:-default}"
 export INSTALL_BOOKINFO="${INSTALL_BOOKINFO:-true}"
 export SLEEP_TIME="${SLEEP_TIME:-30}"
+export REMOVE_BINS="${REMOVE_BINS:-false}"
 
 # Check for Root user.
 if [ "$(id -u)" != "0" ]; then
@@ -25,7 +26,7 @@ if [ "$(id -u)" != "0" ]; then
 fi
 
 # Check for kubectl config file.
-if ! [ "$(stat ${KUBECONFIG})" ] ; then
+if ! [ "$(stat ${KUBECONFIG} 2> /dev/null)" ] ; then
     echo "### You must store your tenant cluster credentials to ${KUBECONFIG} before running this script."
     exit 1
 fi
@@ -34,60 +35,47 @@ fi
 OS="$(uname)"
 if [ "x${OS}" = "xDarwin" ] ; then
   OSEXT="osx"
+  KOSEXT="darwin"
 else
   OSEXT="linux"
-fi
-
-# Download the latest version of Istio if ISTIO_VERSION is not specified.
-if [ "x${ISTIO_VERSION}" = "x" ] ; then
-  ISTIO_VERSION=$(curl -sL https://api.github.com/repos/istio/istio/releases/latest | \
-                  grep tag_name | sed "s/ *\"tag_name\": *\"\(.*\)\",*/\1/")
-fi
-
-# Check for existence of istioctl and project directory.
-ISTIOCTL_CHECK="$(istioctl version | grep ${ISTIO_VERSION})"
-if [ "${ISTIOCTL_CHECK}" ] ; then
-    echo "### istioctl ${ISTIO_VERSION} currently installed, continuing with clean-up ..."
-else
-    echo "### istioctl ${ISTIO_VERSION} not installed, exiting clean-up ..."
-    exit 1
+  KOSEXT="linux"
 fi
 
 # kubectl binary download and setup.
-KUBECTL_CHECK="$(kubectl version --client --short | grep v${KUBECTL_VERSION})"
+KUBECTL_CHECK="$(kubectl version --client --short 2> /dev/null | grep v${KUBECTL_VERSION})"
 if [ "${KUBECTL_CHECK}" ] ; then
-    echo "### kubectl ${KUBECTL_VERSION} currently installed, continuing with clean-up ..."
+    echo "### kubectl v${KUBECTL_VERSION} currently installed, continuing with clean-up ..."
 else
-    echo "### kubectl v${KUBECTL_VERSION} not installed, exiting clean-up ..."
-    exit 1
-fi
-
-# helm client binary download and setup.
-HELM_CHECK="$(helm version --client --short | grep v${HELM_VERSION})"
-if [ "${HELM_CHECK}" ] ; then
-    echo "### helm v${HELM_VERSION} currently installed, continuing with clean-up ..."
-else
-    echo "### helm v${HELM_VERSION} not installed, exiting clean-up  ..."
-    exit 1
+    echo "### kubectl v${KUBECTL_VERSION} is required for clean-up, installing kubectl ..."
+    # kubectl binary download and setup.
+    NAME="kubectl"
+    URL="https://storage.googleapis.com/kubernetes-release/release/v${KUBECTL_VERSION}/bin/${KOSEXT}/amd64/kubectl"
+    if ! [ "$(stat ${NAME} 2> /dev/null)" ] ; then
+        echo "### Downloading ${NAME} v${KUBECTL_VERSION} from ${URL} ..."
+        curl -sLO "${URL}"
+    fi
+    # Move kubectl binary to ${BIN_DIR}
+    chmod +x ./${NAME}
+    mv ./${NAME} ${BIN_DIR}
+    echo "### ${NAME} v${KUBECTL_VERSION} binary installed at ${BIN_DIR} ..."
 fi
 
 # Remove bookinfo sample app
 if [ "${INSTALL_BOOKINFO}" = "true" ] ; then
-    kubectl get po | grep productpage
+    kubectl get po 2> /dev/null | grep productpage
     if [ $? -eq 0 ] ; then
         echo "Deleting bookinfo deployment"
         kubectl delete -f istio-${ISTIO_VERSION}/samples/bookinfo/kube/bookinfo.yaml
     fi
-    kubectl get ing | grep gateway
+    kubectl get ing 2> /dev/null | grep gateway
     if [ $? -eq 0 ] ; then
         echo "Deleting bookinfo ingress"
         kubectl delete -f istio-${ISTIO_VERSION}/samples/bookinfo/kube/bookinfo-gateway.yaml
     fi
 fi   
 
-
-# Remove Istio control-plane deployment.
-kubectl get deploy -n ${ISTIO_NAMESPACE} | grep istio-pilot
+# Remove Istio deployment.
+kubectl get deploy -n ${ISTIO_NAMESPACE} 2> /dev/null | grep istio-pilot
 if [ $? -eq 0 ] ; then
     echo "### Deleting Kubernetes deployment for Istio in namespace \"${ISTIO_NAMESPACE}\" ..."
     kubectl delete -f $HOME/istio-${ISTIO_VERSION}.yaml
@@ -99,7 +87,7 @@ else
 fi
 
 # Remove ISTIO_INJECT_NS namespace label used for Istio automatic sidecar injection.
-kubectl get namespace -L istio-injection | grep ${ISTIO_INJECT_NS} | grep enabled
+kubectl get namespace -L istio-injection 2> /dev/null | grep ${ISTIO_INJECT_NS} | grep enabled
 if [ $? -eq 0 ] ; then
     echo "### Removing \"istio-injection=enabled\" from \"${ISTIO_INJECT_NS}\" namespace ..."
     kubectl label namespace ${ISTIO_INJECT_NS} istio-injection-
@@ -107,8 +95,8 @@ else
     echo "### Label \"istio-injection=enabled\" does not exist for \"${ISTIO_INJECT_NS}\", skipping ..."
 fi
 
-# Remove Kubernetes ISTIO_NAMESPACE namespace used by Istio control-plane.
-kubectl get ns ${ISTIO_NAMESPACE}
+# Remove Kubernetes ISTIO_NAMESPACE namespace used by Istio.
+kubectl get ns ${ISTIO_NAMESPACE} 2> /dev/null
 if [ $? -eq 0 ] ; then
     echo "### Removing \"${ISTIO_NAMESPACE}\" namespace ..."
     kubectl delete ns ${ISTIO_NAMESPACE}
@@ -118,43 +106,46 @@ fi
 
 # Render Kubernetes manifest for Istio deployment.
 MANIFEST="$HOME/istio-${ISTIO_VERSION}.yaml"
-if [ "$(stat ${MANIFEST})" ]; then
+if [ "$(stat ${MANIFEST} 2> /dev/null)" ]; then
     echo "### Removing manifest ${MANIFEST} ..."
     rm -rf $MANIFEST
 else
-    echo "### Kubernetes manifest ${MANIFEST} cdoes not exist, skipping ..."
-fi
-
-# Remove helm client binary.
-if [ "${HELM_CHECK}" ] ; then
-    echo "### Removing helm v${HELM_VERSION} from ${BIN_DIR} ..."
-    rm -rf ${BIN_DIR}/helm
-else
-    echo "### helm v${HELM_VERSION} not installed in ${BIN_DIR}, skipping ..."
-fi
-
-# Remove kubectl client binary.
-if [ "${KUBECTL_CHECK}" ] ; then
-    echo "### Removing kubectl v${KUBECTL_VERSION} from ${BIN_DIR} ..."
-    rm -rf ${BIN_DIR}/kubectl
-else
-    echo "### kubectl v${KUBECTL_VERSION} not installed in ${BIN_DIR}, skipping ..."
-fi
-
-# Remove istioctl client binary.
-if [ i"${ISTIOCTL_CHECK}" ] ; then
-    echo "### Removing istioctl v${ISTIO_VERSION} from ${BIN_DIR} ..."
-    rm -rf ${BIN_DIR}/istioctl
-else
-    echo "### istioctl v${ISTIO_VERSION} not installed in ${BIN_DIR}, skipping ..."
+    echo "### Kubernetes manifest ${MANIFEST} does not exist, skipping ..."
 fi
 
 # Remove Istio project directory
-if [ "$(stat istio-${ISTIO_VERSION})" ] ; then
+if [ "$(stat istio-${ISTIO_VERSION} 2> /dev/null)" ] ; then
     echo "### Removing istio-${ISTIO_VERSION} project directory ..."
     rm -rf istio-${ISTIO_VERSION}
 else
     echo "### Istio project directory \"istio-${ISTIO_VERSION}\" does not exist, skipping ..."
 fi
 
-echo "### Istio installation clean-up complete!"
+# Remove binaries.
+if [ "${REMOVE_BINS}" = "true" ] ; then
+    # Remove helm client binary.
+    HELM_CHECK="$(helm version --client --short 2> /dev/null | grep v${HELM_VERSION})"
+    if [ "${HELM_CHECK}" ] ; then
+        echo "### Removing helm v${HELM_VERSION} from ${BIN_DIR} ..."
+        rm -rf ${BIN_DIR}/helm
+    else
+        echo "### helm v${HELM_VERSION} not installed in ${BIN_DIR}, skipping ..."
+    fi
+    # Remove kubectl client binary.
+    if [ "${KUBECTL_CHECK}" ] ; then
+        echo "### Removing kubectl v${KUBECTL_VERSION} from ${BIN_DIR} ..."
+        rm -rf ${BIN_DIR}/kubectl
+    else
+        echo "### kubectl v${KUBECTL_VERSION} not installed in ${BIN_DIR}, skipping ..."
+    fi
+    # Remove istioctl client binary.
+    ISTIOCTL_CHECK="$(istioctl version 2> /dev/null | grep ${ISTIO_VERSION})"
+    if [ "${ISTIOCTL_CHECK}" ] ; then
+        echo "### Removing istioctl v${ISTIO_VERSION} from ${BIN_DIR} ..."
+        rm -rf ${BIN_DIR}/istioctl
+    else
+        echo "### istioctl v${ISTIO_VERSION} not installed in ${BIN_DIR}, skipping ..."
+    fi
+fi
+
+echo "### Istio clean-up complete!"
